Use the JoinHandle to ensure worker is cleaned up
Hi @matklad !  I've been using this pattern for a few projects, but have recently run in to a bug...  I hadn't fully appreciated the implications of dropping the `JoinHandle` returned by `thread::spawn` like these examples all do, in particular that variables in the worker might not be cleanly dropped when the main program exits.  See https://ianrrees.github.io/2022/07/29/stopping-rust-threads.html for a poorly written and formatted blog post.

It's interesting to compare Rust's behaviour with C++ in handling detached threads.  I made a quick test with gcc 11 and C++11, it throws on exiting main when the detached worker thread is still running - one could argue that's more Rusty than Rust!

I've only modified the post's canonical example in this PR, because I'm not sure how the `.join()` fits in with your exposition style, plus I'm not familiar enough with async to modify that example.

Thanks for doing the things you do!  I'm a big fan of rust-analyzer.